change home logic
